refactor content schemas into separat folders
