+++ packages/lightnet/__tests__/content/query-media-items.spec.ts
@@ -1,10 +1,8 @@
 import { expect, test, vi } from "vitest"
 
-import type {
-  MediaCollectionEntry,
-  MediaItemEntry,
-} from "../../src/content/content-schema"
 import { queryMediaItems } from "../../src/content/query-media-items"
+import type { MediaCollectionEntry } from "../../src/content/schema/media-collection"
+import type { MediaItemEntry } from "../../src/content/schema/media-item"
 
 vi.mock("astro:content", async () => {
   const { z } = await import("astro/zod")
