+++ packages/lightnet/src/content/schema/category.ts
@@ -1,0 +1,39 @@
+import { z } from "astro/zod"
+import type { SchemaContext } from "astro:content"
+
+import { imageSchema } from "../astro-image"
+import { inlineTranslationSchema } from "./inline-translation"
+
+/**
+ * Category Schema
+ */
+export const categorySchema = z.object({
+  /**
+   * Name of the category.
+   *
+   * Label translated for the default locale. Other configured site locales are optional.
+   */
+  label: inlineTranslationSchema,
+
+  /* Relative path to the thumbnail image of this category.
+   *
+   * The image is expected to be inside the `images` folder next to category definition json.
+   * It can have one of these file types: png, jpg, tiff, webp, gif, svg, avif.
+   * We suggest to give it a size of at least 1000px for it's longer side.
+   *
+   * @example "./images/devotionals.jpg"
+   */
+  image: imageSchema.optional(),
+})
+
+export const createCategorySchema = ({ image }: SchemaContext) =>
+  categorySchema.extend({
+    image: image().optional(),
+  })
+
+export const categoryEntrySchema = z.object({
+  id: z.string(),
+  data: categorySchema,
+})
+
+export type CategoryEntry = z.infer<typeof categoryEntrySchema>
