+++ packages/lightnet/src/content/schema/media-collection.ts
@@ -1,0 +1,30 @@
+import { z } from "astro/zod"
+import { reference } from "astro:content"
+
+import { inlineTranslationSchema } from "./inline-translation"
+
+/**
+ * Media Collection Schema
+ */
+export const mediaCollectionSchema = z.object({
+  /**
+   * Name of the collection.
+   *
+   * Label translated for the default locale. Other configured site locales are optional.
+   */
+  label: inlineTranslationSchema,
+  /**
+   * Ordered list of media items included in this collection.
+   * The array order defines how items are shown when querying by collection.
+   *
+   * @example ["my-book--en", "my-video--en"]
+   */
+  mediaItems: z.array(reference("media")),
+})
+
+export const mediaCollectionEntrySchema = z.object({
+  id: z.string(),
+  data: mediaCollectionSchema,
+})
+
+export type MediaCollectionEntry = z.infer<typeof mediaCollectionEntrySchema>
